roup::ir

Enum ExpressionKind

pub enum ExpressionKind<'a> {
    IntLiteral(i64),
    Identifier(&'a str),
    BinaryOp {
        left: Box<ExpressionAst<'a>>,
        op: BinaryOperator,
        right: Box<ExpressionAst<'a>>,
    },
    UnaryOp {
        op: UnaryOperator,
        operand: Box<ExpressionAst<'a>>,
    },
    Call {
        function: &'a str,
        args: Vec<ExpressionAst<'a>>,
    },
    ArrayAccess {
        array: Box<ExpressionAst<'a>>,
        indices: Vec<ExpressionAst<'a>>,
    },
    Conditional {
        condition: Box<ExpressionAst<'a>>,
        then_expr: Box<ExpressionAst<'a>>,
        else_expr: Box<ExpressionAst<'a>>,
    },
    Parenthesized(Box<ExpressionAst<'a>>),
    Complex(&'a str),
}

Common expression patterns in OpenMP directives

Learning: Large Enums with Data

This enum demonstrates Rustโs powerful enum system. Each variant can carry different data:

IntLiteral(i64) - carries an integer

Identifier(&str) - carries a string reference

BinaryOp { ... } - carries multiple fields

This is much more powerful than C enums, which can only be simple tags.

Variants

IntLiteral(i64)

Integer literal: 42, 0x10, 0b1010

Identifier(&'a str)

Identifier: N, num_threads, my_var

BinaryOp

Binary operation: a + b, N * 2, i < 10

Fields

left: Box<ExpressionAst<'a>>

op: BinaryOperator

right: Box<ExpressionAst<'a>>

UnaryOp

Unary operation: -x, !flag, *ptr

Fields

op: UnaryOperator

operand: Box<ExpressionAst<'a>>

Call

Function call: foo(a, b), omp_get_num_threads()

Fields

function: &'a str

args: Vec<ExpressionAst<'a>>

ArrayAccess

Array subscript: arr[i], matrix[i][j]

Fields

array: Box<ExpressionAst<'a>>

indices: Vec<ExpressionAst<'a>>

Conditional

Ternary conditional: cond ? a : b

Fields

condition: Box<ExpressionAst<'a>>

then_expr: Box<ExpressionAst<'a>>

else_expr: Box<ExpressionAst<'a>>

Parenthesized(Box<ExpressionAst<'a>>)

Parenthesized: (expr)

Complex(&'a str)

Too complex to parse, kept as string

This is our escape hatch for expressions that are valid but not yet supported by the parser.