roup::ir

Enum DirectiveKind

#[repr(C)]
pub enum DirectiveKind {

    Parallel = 0,
    ParallelFor = 1,
    ParallelForSimd = 2,
    ParallelSections = 3,
    ParallelWorkshare = 4,
    ParallelLoop = 5,
    ParallelLoopSimd = 8,
    ParallelMasked = 6,
    ParallelMaskedTaskloop = 9,
    ParallelMaskedTaskloopSimd = 17,
    ParallelMaster = 7,
    ParallelMasterTaskloop = 18,
    ParallelMasterTaskloopSimd = 19,
    For = 10,
    ForSimd = 11,
    Sections = 12,
    Section = 13,
    Single = 14,
    Workshare = 15,
    Loop = 16,
    Simd = 20,
    DeclareSimd = 21,
    Task = 30,
    Taskloop = 31,
    TaskloopSimd = 32,
    Taskyield = 33,
    Taskwait = 34,
    Taskgroup = 35,
    Taskgraph = 36,
    TaskIteration = 37,
    MaskedTaskloop = 38,
    MaskedTaskloopSimd = 39,
    Target = 40,
    TargetData = 41,
    TargetEnterData = 42,
    TargetExitData = 43,
    TargetUpdate = 44,
    TargetParallel = 45,
    TargetParallelFor = 46,
    TargetParallelForSimd = 47,
    TargetParallelLoop = 48,
    TargetParallelLoopSimd = 56,
    TargetSimd = 49,
    TargetLoop = 57,
    TargetLoopSimd = 58,
    TargetTeams = 50,
    TargetTeamsDistribute = 51,
    TargetTeamsDistributeSimd = 52,
    TargetTeamsDistributeParallelFor = 53,
    TargetTeamsDistributeParallelForSimd = 54,
    TargetTeamsDistributeParallelLoop = 59,
    TargetTeamsDistributeParallelLoopSimd = 69,
    TargetTeamsLoop = 55,
    TargetTeamsLoopSimd = 85,
    EndTarget = 147,
    Teams = 60,
    TeamsDistribute = 61,
    TeamsDistributeSimd = 62,
    TeamsDistributeParallelFor = 63,
    TeamsDistributeParallelForSimd = 64,
    TeamsDistributeParallelLoop = 66,
    TeamsDistributeParallelLoopSimd = 67,
    TeamsLoop = 65,
    TeamsLoopSimd = 68,
    Barrier = 70,
    Critical = 71,
    Atomic = 72,
    AtomicRead = 77,
    AtomicWrite = 78,
    AtomicUpdate = 79,
    AtomicCapture = 86,
    AtomicCompareCapture = 87,
    Flush = 73,
    Ordered = 74,
    Master = 75,
    Masked = 76,
    DeclareReduction = 80,
    DeclareMapper = 81,
    DeclareTarget = 82,
    DeclareVariant = 83,
    DeclareInduction = 84,
    Distribute = 90,
    DistributeSimd = 91,
    DistributeParallelFor = 92,
    DistributeParallelForSimd = 93,
    DistributeParallelLoop = 94,
    DistributeParallelLoopSimd = 95,
    Metadirective = 100,
    BeginMetadirective = 103,
    Assume = 101,
    Assumes = 102,
    BeginAssumes = 104,
    BeginDeclareTarget = 112,
    EndDeclareTarget = 113,
    BeginDeclareVariant = 114,
    EndDeclareVariant = 115,
    Tile = 105,
    Unroll = 106,
    Fuse = 107,
    Split = 108,
    Interchange = 109,
    Reverse = 110,
    Stripe = 111,
    Threadprivate = 120,
    Allocate = 121,
    Allocators = 122,
    Requires = 123,
    Scan = 124,
    Depobj = 125,
    Nothing = 126,
    Error = 127,
    Cancel = 128,
    CancellationPoint = 129,
    Dispatch = 130,
    Interop = 131,
    Scope = 132,
    Groupprivate = 133,
    Workdistribute = 134,
    Do = 135,
    DoSimd = 136,
    ParallelDo = 137,
    ParallelDoSimd = 138,
    DistributeParallelDo = 139,
    DistributeParallelDoSimd = 140,
    TeamsDistributeParallelDo = 141,
    TeamsDistributeParallelDoSimd = 142,
    TargetParallelDo = 143,
    TargetParallelDoSimd = 144,
    TargetTeamsDistributeParallelDo = 145,
    TargetTeamsDistributeParallelDoSimd = 146,
    Unknown = 255,

}

OpenMP directive type

This enum covers all standard OpenMP directives from the 5.2 specification. Each directive is represented as a unique variant.

Examples

let kind = DirectiveKind::Parallel;
assert_eq!(kind.to_string(), "parallel");

let kind = DirectiveKind::ParallelFor;
assert_eq!(kind.to_string(), "parallel for");

Learning: Large Enum with Clear Organization

This enum demonstrates organizing a large number of variants (70+) into logical categories using comments and grouping.

Variants

Parallel = 0

#pragma omp parallel

ParallelFor = 1

#pragma omp parallel for

ParallelForSimd = 2

#pragma omp parallel for simd

ParallelSections = 3

#pragma omp parallel sections

ParallelWorkshare = 4

#pragma omp parallel workshare (Fortran)

ParallelLoop = 5

#pragma omp parallel loop

ParallelLoopSimd = 8

#pragma omp parallel loop simd

ParallelMasked = 6

#pragma omp parallel masked

ParallelMaskedTaskloop = 9

#pragma omp parallel masked taskloop

ParallelMaskedTaskloopSimd = 17

#pragma omp parallel masked taskloop simd

ParallelMaster = 7

#pragma omp parallel master (deprecated in 5.1)

ParallelMasterTaskloop = 18

#pragma omp parallel master taskloop

ParallelMasterTaskloopSimd = 19

#pragma omp parallel master taskloop simd

For = 10

#pragma omp for

ForSimd = 11

#pragma omp for simd

Sections = 12

#pragma omp sections

Section = 13

#pragma omp section

Single = 14

#pragma omp single

Workshare = 15

#pragma omp workshare (Fortran)

Loop = 16

#pragma omp loop

Simd = 20

#pragma omp simd

DeclareSimd = 21

#pragma omp declare simd

Task = 30

#pragma omp task

Taskloop = 31

#pragma omp taskloop

TaskloopSimd = 32

#pragma omp taskloop simd

Taskyield = 33

#pragma omp taskyield

Taskwait = 34

#pragma omp taskwait

Taskgroup = 35

#pragma omp taskgroup

Taskgraph = 36

#pragma omp taskgraph (OpenMP 6.0)

TaskIteration = 37

#pragma omp task iteration (OpenMP 6.0)

MaskedTaskloop = 38

#pragma omp masked taskloop

MaskedTaskloopSimd = 39

#pragma omp masked taskloop simd

Target = 40

#pragma omp target

TargetData = 41

#pragma omp target data

TargetEnterData = 42

#pragma omp target enter data

TargetExitData = 43

#pragma omp target exit data

TargetUpdate = 44

#pragma omp target update

TargetParallel = 45

#pragma omp target parallel

TargetParallelFor = 46

#pragma omp target parallel for

TargetParallelForSimd = 47

#pragma omp target parallel for simd

TargetParallelLoop = 48

#pragma omp target parallel loop

TargetParallelLoopSimd = 56

#pragma omp target parallel loop simd

TargetSimd = 49

#pragma omp target simd

TargetLoop = 57

#pragma omp target loop

TargetLoopSimd = 58

#pragma omp target loop simd

TargetTeams = 50

#pragma omp target teams

TargetTeamsDistribute = 51

#pragma omp target teams distribute

TargetTeamsDistributeSimd = 52

#pragma omp target teams distribute simd

TargetTeamsDistributeParallelFor = 53

#pragma omp target teams distribute parallel for

TargetTeamsDistributeParallelForSimd = 54

#pragma omp target teams distribute parallel for simd

TargetTeamsDistributeParallelLoop = 59

#pragma omp target teams distribute parallel loop

TargetTeamsDistributeParallelLoopSimd = 69

#pragma omp target teams distribute parallel loop simd

TargetTeamsLoop = 55

#pragma omp target teams loop

TargetTeamsLoopSimd = 85

#pragma omp target teams loop simd

EndTarget = 147

#pragma omp end target

Teams = 60

#pragma omp teams

TeamsDistribute = 61

#pragma omp teams distribute

TeamsDistributeSimd = 62

#pragma omp teams distribute simd

TeamsDistributeParallelFor = 63

#pragma omp teams distribute parallel for

TeamsDistributeParallelForSimd = 64

#pragma omp teams distribute parallel for simd

TeamsDistributeParallelLoop = 66

#pragma omp teams distribute parallel loop

TeamsDistributeParallelLoopSimd = 67

#pragma omp teams distribute parallel loop simd

TeamsLoop = 65

#pragma omp teams loop

TeamsLoopSimd = 68

#pragma omp teams loop simd

Barrier = 70

#pragma omp barrier

Critical = 71

#pragma omp critical

Atomic = 72

#pragma omp atomic

AtomicRead = 77

#pragma omp atomic read

AtomicWrite = 78

#pragma omp atomic write

AtomicUpdate = 79

#pragma omp atomic update

AtomicCapture = 86

#pragma omp atomic capture

AtomicCompareCapture = 87

#pragma omp atomic compare capture

Flush = 73

#pragma omp flush

Ordered = 74

#pragma omp ordered

Master = 75

#pragma omp master

Masked = 76

#pragma omp masked

DeclareReduction = 80

#pragma omp declare reduction

DeclareMapper = 81

#pragma omp declare mapper

DeclareTarget = 82

#pragma omp declare target

DeclareVariant = 83

#pragma omp declare variant

DeclareInduction = 84

#pragma omp declare induction (OpenMP 6.0)

Distribute = 90

#pragma omp distribute

DistributeSimd = 91

#pragma omp distribute simd

DistributeParallelFor = 92

#pragma omp distribute parallel for

DistributeParallelForSimd = 93

#pragma omp distribute parallel for simd

DistributeParallelLoop = 94

#pragma omp distribute parallel loop

DistributeParallelLoopSimd = 95

#pragma omp distribute parallel loop simd

Metadirective = 100

#pragma omp metadirective

BeginMetadirective = 103

#pragma omp begin metadirective

Assume = 101

#pragma omp assume (OpenMP 6.0)

Assumes = 102

#pragma omp assumes (OpenMP 6.0)

BeginAssumes = 104

#pragma omp begin assumes

BeginDeclareTarget = 112

#pragma omp begin declare target

EndDeclareTarget = 113

#pragma omp end declare target

BeginDeclareVariant = 114

#pragma omp begin declare variant

EndDeclareVariant = 115

#pragma omp end declare variant

Tile = 105

#pragma omp tile

Unroll = 106

#pragma omp unroll

Fuse = 107

#pragma omp fuse

Split = 108

#pragma omp split

Interchange = 109

#pragma omp interchange

Reverse = 110

#pragma omp reverse

Stripe = 111

#pragma omp stripe

Threadprivate = 120

#pragma omp threadprivate

Allocate = 121

#pragma omp allocate

Allocators = 122

#pragma omp allocators (OpenMP 6.0)

Requires = 123

#pragma omp requires

Scan = 124

#pragma omp scan

Depobj = 125

#pragma omp depobj

Nothing = 126

#pragma omp nothing

Error = 127

#pragma omp error

Cancel = 128

#pragma omp cancel (OpenMP 4.0)

CancellationPoint = 129

#pragma omp cancellation point (OpenMP 4.0)

Dispatch = 130

#pragma omp dispatch (OpenMP 6.0)

Interop = 131

#pragma omp interop (OpenMP 5.1)

Scope = 132

#pragma omp scope (OpenMP 5.1)

Groupprivate = 133

#pragma omp groupprivate (OpenMP 6.0)

Workdistribute = 134

#pragma omp workdistribute (Fortran, OpenMP 6.0)

Do = 135

!$omp do (Fortran equivalent of for)

DoSimd = 136

!$omp do simd (Fortran equivalent of for simd)

ParallelDo = 137

!$omp parallel do (Fortran equivalent of parallel for)

ParallelDoSimd = 138

!$omp parallel do simd (Fortran equivalent of parallel for simd)

DistributeParallelDo = 139

!$omp distribute parallel do (Fortran)

DistributeParallelDoSimd = 140

!$omp distribute parallel do simd (Fortran)

TeamsDistributeParallelDo = 141

!$omp teams distribute parallel do (Fortran)

TeamsDistributeParallelDoSimd = 142

!$omp teams distribute parallel do simd (Fortran)

TargetParallelDo = 143

!$omp target parallel do (Fortran)

TargetParallelDoSimd = 144

!$omp target parallel do simd (Fortran)

TargetTeamsDistributeParallelDo = 145

!$omp target teams distribute parallel do (Fortran)

TargetTeamsDistributeParallelDoSimd = 146

!$omp target teams distribute parallel do simd (Fortran)

Unknown = 255

Unknown or custom directive

Implementations

impl DirectiveKind

pub const fn name_for_language(self, language: Language) -> &'static str

Get the language-specific directive name

Returns the appropriate directive name for the given language context. For C/C++, returns forms like “for”, “parallel for”. For Fortran, returns forms like “do”, “parallel do”.

This method is used for cross-language translation of directives.

Examples

assert_eq!(DirectiveKind::For.name_for_language(Language::C), "for");
assert_eq!(DirectiveKind::For.name_for_language(Language::Fortran), "do");
assert_eq!(DirectiveKind::ParallelFor.name_for_language(Language::C), "parallel for");
assert_eq!(DirectiveKind::ParallelFor.name_for_language(Language::Fortran), "parallel do");

pub const fn as_str(self) -> &'static str

Get the canonical directive name as a string

Returns the standard C/C++ form of the directive name. This is the same as the Display implementation.

Examples

assert_eq!(DirectiveKind::Parallel.as_str(), "parallel");
assert_eq!(DirectiveKind::ParallelFor.as_str(), "parallel for");

pub fn is_parallel(&self) -> bool

Check if this is a parallel construct

pub fn is_worksharing(&self) -> bool

Check if this is a work-sharing construct

pub fn is_simd(&self) -> bool

Check if this is a SIMD construct

pub fn is_task(&self) -> bool

Check if this is a task construct

pub fn is_target(&self) -> bool

Check if this is a target construct

pub fn is_teams(&self) -> bool

Check if this is a teams construct

pub fn is_loop(&self) -> bool

Check if this is a loop construct

pub fn is_synchronization(&self) -> bool

Check if this is a synchronization construct

pub fn is_declare(&self) -> bool

Check if this is a declare construct

pub fn has_structured_block(&self) -> bool

Check if this directive has a structured block (requires end directive)